refactor:clean up posthog init component
